Move TyBuilder to its own module
